requirements.txt
flask==3.0.2
torch==2.2.0
transformers==4.38.2
pillow==10.2.0
numpy==1.26.4
huggingface_hub==0.21.4 